+++ appserver/web/web-core/src/main/java/org/apache/catalina/core/StandardServer.java
@@ -62,7 +62,7 @@ package org.apache.catalina.core;
 import org.apache.catalina.*;
 import org.apache.catalina.deploy.NamingResources;
 import org.apache.catalina.util.LifecycleSupport;
-import org.apache.catalina.util.LogCleanerUtil;
+import com.sun.logging.LogCleanerUtil;
 
 import javax.management.ObjectName;
 import java.beans.PropertyChangeListener;
