+++ appserver/web/web-core/src/main/java/org/apache/catalina/realm/RealmBase.java
@@ -58,6 +58,7 @@
 
 package org.apache.catalina.realm;
 
+import com.sun.logging.LogCleanerUtil;
 import org.apache.catalina.*;
 import org.apache.catalina.authenticator.AuthenticatorBase;
 import org.apache.catalina.connector.Response;
