+++ nucleus/common/common-util/pom.xml
@@ -75,8 +75,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                     <excludes>
                         <exclude>**/.ade_path/**</exclude>
                         <exclude>**/ha/spi/*.java</exclude>
